requirements.txt
Django==3.0
django-crispy-forms==1.8.1
django-bootstrap-datepicker-plus==3.0.5